services/camera-service/requirements.txt
opencv-python-headless
numpy
pydantic
python-dotenv
grpcio
grpcio-tools
